% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/any_na.R
\name{any_na}
\alias{any_na}
\alias{all_na}
\alias{non_na}
\title{Any NA? (computed in a vectorized manner)}
\usage{
any_na(...)

all_na(...)

non_na(...)
}
\description{
Any NA? (computed in a vectorized manner)

All NA? (computed in a vectorized manner)

The first non-na element (computed in a vectorized manner)
}
