% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conform_dt.R
\name{conform_dt}
\alias{conform_dt}
\title{Make All groups the Same Shape}
\usage{
conform_dt(x, group.by, index.by, reorder.cols = TRUE)
}
\arguments{
\item{x}{a data.frame/data.table object}

\item{group.by}{column names or column positions}

\item{index.by}{column names or column positions}

\item{reorder.cols}{}
}
\value{
a data.table object
}
\description{
Make All groups the Same Shape
}
