% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{exp_moving_avg}
\alias{exp_moving_avg}
\title{Running Exponential Moving Average (EMA)}
\usage{
exp_moving_avg(x, decay, min_size = 1L)
}
\arguments{
\item{x}{numeric vector}

\item{decay}{double, the decay factor}

\item{min_size}{if number of non-NA elements is fewer than min_size, return NA. For EMA,
this parameter is only used to remove the first a few elements.}
}
\description{
Missing Values are forward filled.
}
