% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ffill}
\alias{ffill}
\title{Forward Fill NAs}
\usage{
ffill(x, w = 0L)
}
\arguments{
\item{x}{numeric vector}

\item{w}{integer, windows size, results are right aligned if w < 0, left aligned if w > 0}
}
\description{
Forward Fill NAs
}
