% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_except.R
\name{rm_except}
\alias{rm_except}
\title{Remove all variables except ones specified}
\usage{
rm_except(..., list = character(), all = FALSE)
}
\arguments{
\item{...}{the objects to be removed, as names (unquoted) or character strings (quoted)}

\item{list}{a character vector naming objects to be kept}

\item{all}{will be paased to the \code{ls()} function}
}
\description{
Remove all variables except ones specified
}
