% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{running_mean}
\alias{running_mean}
\title{Running mean
                                                                          ../////}
\usage{
running_mean(x, w, min_size = 1L)
}
\arguments{
\item{x}{numeric vector}

\item{w}{integer, windows size, results are right aligned if w < 0, left aligned if w > 0}

\item{min_size}{if number of non-NA elements is fewer than min_size, return NA}
}
\description{
Running mean
                                                                          ../////
}
