% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{running_min}
\alias{running_min}
\title{Running Min}
\usage{
running_min(x, w, min_size = 1L)
}
\arguments{
\item{x}{numeric vector}

\item{w}{integer, windows size, results are right aligned if w < 0, left aligned if w > 0}
}
\description{
This O(N) time solution is borrowed from Stackoverflow, I think it is 
easier to implement than the two array solution
}
\details{
at every step:

if (!Deque.Empty) and (Deque.Head.Index <= CurrentIndex - T) then 
     Deque.ExtractHead;
  //Head is too old, it is leaving the window

while (!Deque.Empty) and (Deque.Tail.Value > CurrentValue) do
     Deque.ExtractTail;
  //remove elements that have no chance to become minimum in the window

Deque.AddTail(CurrentValue, CurrentIndex); 
  CurrentMin = Deque.Head.Value
  //Head value is minimum in the current window
}
