% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{running_sd}
\alias{running_sd}
\title{Running Standard Deviation}
\usage{
running_sd(x, w, min_size = 2L)
}
\arguments{
\item{x}{numeric vector}

\item{w}{integer, windows size, results are right aligned if w < 0, left aligned if w > 0}

\item{min_size}{if number of non-NA elements is fewer than min_size, return NA}
}
\description{
Running Standard Deviation
}
