% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{running_sd0}
\alias{running_sd0}
\title{Running Standard Deviation assuming Mean = 0}
\usage{
running_sd0(x, w, min_size = 2L)
}
\arguments{
\item{x}{numeric vector}

\item{w}{integer, windows size, results are right aligned if w < 0, left aligned if w > 0}

\item{min_size}{if number of non-NA elements is fewer than min_size, return NA}
}
\description{
Running Standard Deviation assuming Mean = 0
}
